refactor: colocates reader layout with it's dependencies
