+++ src/app/read/layout.tsx
@@ -1,1 +1,26 @@
-export { ReaderLayout as default } from '~/layouts'
+import type { ReactNode } from 'react'
+
+import { getBookListWithCache } from '~/db'
+import { BottomToolbar, TopToolbar, VerseDetailsMenuRoot } from '~/organisms'
+
+import { Footer } from './lib/Footer'
+import { UseRestorePrevSessionChapter } from './lib/UseRestorePrevSessionChapter'
+import { UseSetUpReaderState } from './lib/UseSetUpReaderState'
+
+const ReaderLayout = async ({ children }: { children: ReactNode }) => {
+	const bookList = await getBookListWithCache()
+
+	return (
+		<>
+			<UseRestorePrevSessionChapter />
+			<UseSetUpReaderState bookList={bookList} />
+			<TopToolbar />
+			{children}
+			<BottomToolbar />
+			<VerseDetailsMenuRoot bookList={bookList} />
+			<Footer />
+		</>
+	)
+}
+
+export default ReaderLayout
