+++ src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './buildReaderUrl'
 export * from './useRangeInput'
 export * from './useSelectedLayoutPathname'
 export * from './useSetupClientState'
