--- src/layouts/ReaderLayout/index.ts
@@ -1,1 +1,0 @@
-export * from './Reader.layout'
