+++ src/layouts/index.ts
@@ -1,2 +1,1 @@
-export * from './ReaderLayout'
 export * from './RootLayout'
