+++ src/organisms/BottomToolbar/ReturnFromReferenceFab.tsx
@@ -8,8 +8,8 @@ import { hstack } from 'styled-system/patterns'
 import { button } from 'styled-system/recipes'
 
 import { type TReaderPageParams } from '~/_pages'
+import { buildReaderUrl } from '~/app/read/lib'
 import { Icon } from '~/components'
-import { buildReaderUrl } from '~/hooks'
 import { referenceOriginAtom, selectedReferenceAtom } from '~/state'
 
 const iconCls = css({
